Changed text, layout and button styles
It looks like bootstrap, which it is, but we can make it easier
for people to understand what it does and have consistent calls to
action. Trying to focus more on the problem we try to solve and
not how we solve it, we can put that somewhere else, there may be more of
an opportunity for people who aren't technical to engage with us.